rename and create types in order to use a common type for both match list and tables
